Fix SbsSettings incorrect length error

+++ backend/src/diag/settings.rs
@@ -165,7 +165,7 @@ impl TcuSettings for SolSettings {
 #[derive(Default, Debug, Copy, Clone, PartialEq, PartialOrd, Serialize, Deserialize)]
 #[repr(C, packed)]
 pub  struct SbsSettings {
-    shift_solenoid_pwm_reduction_time: u32,
+    shift_solenoid_pwm_reduction_time: u16,
     delta_rpm_flare_detect: u16,
     f_shown_if_flare: bool,
     torque_request_upshift: bool,
